+++ AU_ERP/AU_ERP/Models/BPTypeNumberRanges.cs
@@ -8,11 +8,11 @@ namespace AU_ERP.Models
         public int RangeID { get; set; }
         public int? BPTypeId { get; set; }
         public string? Prefix { get; set; }
-        public int StartNumber { get; set; }
-        public int EndNumber { get; set; }
+        public long StartNumber { get; set; }
+        public long EndNumber { get; set; }
 
         /// <summary>Last issued number; 0 means next issue will be <see cref="StartNumber"/>.</summary>
-        public int CurrentNumber { get; set; }
+        public long CurrentNumber { get; set; }
 
         public virtual BPTypeSample? TypeSample { get; set; }
     }
